Enhance error logging and handling (#353)
This PR bumps `@grafana/google-sdk` to latest version which includes a
fix for selecting the default authentication type. There is a bug right
now when you create a new data source and doesn't select anything the
authentication type is not filled. This caused nil pointer dereference
in the backend. Added handling in the backend as well. Also refactored
some errors handling code to include context and removed some
meaningless errors for the users.

How to test:
- Create a new data source
- Click Save&test
- Observe plugin error returned

Fixes https://github.com/grafana/google-bigquery-datasource/issues/340