Fix: Panic invalid memory address or nil pointer dereference (#365)
This PR fixes the error `invalid memory address or nil pointer
dereference`. This was caused by running quries through storage api that
were already closed.

To reproduce:

- Create multiple queries
- Enable storage API for them
- Spam the `Run query` button
- Observe the error

The fix is to remove the client.Close call as it deletes the storage api
client and the BigQuery client advises not to call it.

I also changed the error source for more errors.